+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 -
 
+### v1.22.0
+
+- Disable Biconomy for CERE ERC20 transfer
+- Application bundle optimization to improve load speed
+- Add CERE ERC20 token to default assets
+- Disable OPT code auto-capitalization
+- Optimize login flow to decrease time to login
+- Increase assets balance pooling interval to 10 seconds
+
 ### v1.21.0
 
 - Use OpenLogin ID Token as bearer authorization in API requests